requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
PyJWT==2.8.0
Pillow==10.1.0
PyPDF2==3.0.1
pdf2image==1.16.3
reportlab==4.0.7
fillpdf==0.7.2
openai==1.3.5
sentence-transformers==2.2.2
numpy==1.24.3
scikit-learn==1.3.2
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
